organize data format tests
