+++ tests/data/__init__.py
@@ -1,0 +1,121 @@
+import h5py
+import numpy as np
+
+from zea.data.data_format import DatasetElement, generate_zea_dataset
+
+
+def generate_example_dataset(
+    path,
+    add_optional_fields=False,
+    add_optional_dtypes=False,
+    n_frames=2,
+    n_ax=2048,
+    n_el=128,
+    n_tx=11,
+    n_ch=1,
+    sound_speed=1540,
+    center_frequency=7e6,
+    sampling_frequency=40e6,
+    grid_size_z=512,
+    grid_size_x=512,
+):
+    """Generates an example dataset that contains all the necessary fields.
+    Note: This dataset does not contain actual data, but is filled with random
+    values.
+
+    Args:
+        path (str): The path to write the dataset to.
+        add_optional_fields (bool, optional): Whether to add optional fields to
+            the dataset. Defaults to False.
+        add_optional_dtypes (bool, optional): Whether to add optional dtypes to
+            the dataset. Defaults to False.
+    """
+
+    # creating some fake raw and image data
+    raw_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch))
+    # image data is in dB
+    image = np.ones((n_frames, grid_size_z, grid_size_x)) * -40
+
+    # creating some fake scan parameters
+    t0_delays = np.zeros((n_tx, n_el), dtype=np.float32)
+    tx_apodizations = np.zeros((n_tx, n_el), dtype=np.float32)
+    probe_geometry = np.zeros((n_el, 3), dtype=np.float32)
+    probe_geometry[:, 0] = np.linspace(-0.02, 0.02, n_el)
+    initial_times = np.zeros((n_tx,), dtype=np.float32)
+
+    if add_optional_fields:
+        focus_distances = np.ones((n_tx,), dtype=np.float32) * np.inf
+        tx_apodizations = np.zeros((n_tx, n_el), dtype=np.float32)
+        polar_angles = np.zeros((n_tx,), dtype=np.float32)
+        azimuth_angles = np.zeros((n_tx,), dtype=np.float32)
+    else:
+        focus_distances = None
+        tx_apodizations = None
+        polar_angles = None
+        azimuth_angles = None
+
+    if add_optional_dtypes:
+        aligned_data = np.ones((n_frames, n_tx, n_ax, n_el, n_ch))
+        envelope_data = np.ones((n_frames, grid_size_z, grid_size_x))
+        beamformed_data = np.ones((n_frames, grid_size_z, grid_size_x, n_ch))
+        image_sc = np.ones_like(image)
+    else:
+        aligned_data = None
+        envelope_data = None
+        beamformed_data = None
+        image_sc = None
+
+    generate_zea_dataset(
+        path,
+        raw_data=raw_data,
+        aligned_data=aligned_data,
+        envelope_data=envelope_data,
+        beamformed_data=beamformed_data,
+        image=image,
+        image_sc=image_sc,
+        probe_geometry=probe_geometry,
+        sampling_frequency=sampling_frequency,
+        center_frequency=center_frequency,
+        initial_times=initial_times,
+        t0_delays=t0_delays,
+        sound_speed=sound_speed,
+        tx_apodizations=tx_apodizations,
+        probe_name="generic",
+        focus_distances=focus_distances,
+        polar_angles=polar_angles,
+        azimuth_angles=azimuth_angles,
+        additional_elements=_generate_example_dataset_elements(),
+        description="This is an example dataset generated by zea",
+    )
+
+
+def _generate_example_dataset_elements() -> list[DatasetElement]:
+    """Generates a list of example DatasetElement objects to be used as additional
+    elements in the generate_zea_dataset function.
+
+    Returns:
+        list: A list of DatasetElement objects.
+    """
+    example_elements = [
+        DatasetElement(
+            dataset_name="temperature",
+            data=np.array(42),
+            description="The temperature during the measurement",
+            unit="unitless",
+        ),
+        DatasetElement(
+            dataset_name="lens_profile",
+            data=np.random.rand(100),
+            description="An example lens profile",
+            unit="mm",
+            group_name="lens",
+        ),
+        DatasetElement(
+            dataset_name="lens_material",
+            data=np.array(["material1", "material2", "material3"], dtype=h5py.string_dtype()),
+            description="An example lens material list",
+            unit="unitless",
+            group_name="lens",
+        ),
+    ]
+    return example_elements
